Removed custom hook and put normalize func at top of file
