+++ src/components/Forms/ContactForm/ContactForm.js
@@ -2,12 +2,29 @@ import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { useForm } from 'react-hook-form'
 import { FormAlert, onSubmitSuccess, onSubmitError } from '../'
-import { useNormalizePhone } from '../../../hooks'
+
+// Function for input mask on phone number field
+// Formats XXXXXXXXXX as XXX-XXX-XXXX as you type
+const normalizePhone = (value, previousValue) => {
+  if (!value) return value
+
+  const onlyNums = value.replace(/[^\d]/g, '') // only allows 0-9
+
+  if (!previousValue || value.length > previousValue.length) {
+    if (onlyNums.length === 3) return `${onlyNums}`
+    if (onlyNums.length === 6) return `${onlyNums.slice(0, 3)}-${onlyNums.slice(3)}-`
+
+    if (onlyNums.length <= 3) return onlyNums
+    if (onlyNums.length <= 6) return `${onlyNums.slice(0, 3)}-${onlyNums.slice(3)}`
+
+    return `${onlyNums.slice(0, 3)}-${onlyNums.slice(3, 6)}-${onlyNums.slice(6, 10)}`
+  }
+}
 
 function ContactForm() {
+  const [phone, setPhone] = useState('')
   const [loading, setLoading] = useState(false)
   const { register, handleSubmit, errors, reset } = useForm()
-  const { phone, setPhone, onChange } = useNormalizePhone('')
 
   const onSubmit = async (formData, e) => {
     e.preventDefault()
@@ -95,7 +112,10 @@ function ContactForm() {
                 message: 'Please input a valid phone number XXXXXXXXXX',
               },
             })}
-            onChange={e => onChange(e)}
+            onChange={event => {
+              const { value } = event.target
+              setPhone(previousValue => normalizePhone(value, previousValue))
+            }}
             value={phone}
           />
         </div>
