+++ src/hooks/index.js
@@ -1,2 +1,1 @@
 export { default as useScript } from './useScript'
-export { default as useNormalizePhone} from "./useNormalizePhone"
